+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add "--format" flag for CLI commands to choose between rich/plain/grep/json
 - Update info command to provide actual useful information and statistics
 - Default to showing help message without args/command provided
+- Add filter API/CLI command
 
 ## 2026-05-26 - 0.1.2